ASoC: SOF: Add widget list setup and free routines
Add sof_widget_list_setup/free routines that will be used
to dynamically setup and free connected DAPM widgets when
a PCM is opened/closed. These routines accept a list of
connected DAPM widgets as input and determine the SOF
widgets, their corresponding pipeline widgets and connections
between them that need to be setup before the PCM is triggered.
In order to avoid querying the connected DAPM widgets twice
(during hw_params and hw_free), modify the struct snd_sof_pcm_stream
to save the list of widgets and the list of pipelines these widgets
belong to. These will be used to free the widgets and their pipeline
widgets during hw_free.

Add a new field use_count in struct snd_sof_widget which
will be used to keep track of the widget usage. Also, update the
struct snd_sof_route to add the source and sink SOF widgets along
with a field to indicate whether the route has already been setup
or not. This will be used to prevent common routes belonging
to connected DAPM widgets for multiple PCMs from being setup
again.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>