ALSA: hda/hdmi: fix silent stream for first DP connection
When connection type is DisplayPort, the Intel hardware silent streaming
feature does not work as expected for the very first PCM open after display
power-up. Silent streaming is activated, but until at least one real stream
is played, no converter is programmed and with DisplayPort, this results in
audio info frame not to be sent out. As a consequence, the initial samples
of audio playback are lost with receivers that require a preamble before
actual audio is played out.

Fix the issue by properly assigning a converter to the silent stream, and
modifying the sequence to program stream id for this case.

BugLink: https://github.com/thesofproject/linux/issues/2468
Fixes: 951894cf30f41 ("ALSA: hda/hdmi: Add Intel silent stream support")
Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>