ASoC: SOF: Intel: hda: clean up link DMA on prepare for IPC4
In DAI prepare, clean up the link DMA for playback at if the stream
has been stopped since last prepare. This must be done to mirror
the logic of FE PCM prepare, where the pipeleines are reset in this
condition.

Add a new field to struct sof_intel_hda_stream that will be set only in
the case of IPC4.

[combined patches from Ranjani and Peter]
Link: https://github.com/thesofproject/sof/issues/9502
Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>